WEBAPP-240: Test and rename HeaderLanguageSelectorItem
